Remove iter extensions
