--- crates/iter-extensions/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-authors = ["The Nu Project Contributors"]
-description = "Additional iterator functionality"
-edition = "2018"
-license = "MIT"
-name = "iter-extensions"
-version = "0.29.2"
-
-[dependencies]
